@@ -10,7 +10,7 @@
 		"url": "sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=6"
+		"node": ">=8"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -20,7 +20,6 @@
 		"index.d.ts"
 	],
 	"keywords": [
-		"obj",
 		"object",
 		"is",
 		"check",
